feat: Enhance web console with health rings, COBOL ticker, and structured transaction log
Tier 1 — Visual polish:
- Replace plain circle SVG nodes with animated health ring indicators
- Redesign COBOL ticker as a sleek horizontal snippet bar
- Rename "Event Feed" to "Transaction Log" with structured feed items

Tier 2 — Simulation improvements:
- Add day_number and accounts_snapshot to SSE simulation events
- Include per-account balances in simulation state model
- Enhance simulator to track and emit account snapshots each day

Tier 3 — Test alignment:
- Fix 8 E2E Playwright tests for new ticker/modal design
- Add comprehensive simulation unit tests (day progression, SSE format)
- Update CLAUDE.md with current test counts and file references

+++ python/api/models.py
@@ -259,7 +259,7 @@ class SimulationStartRequest(BaseModel):
     """Request body for POST /api/simulation/start."""
     days: int = Field(default=25, gt=0, le=365)    # Number of days to simulate
     seed: Optional[int] = None                      # RNG seed for reproducibility
-    time_scale: int = Field(default=0, ge=0)        # 0=max speed, 3600=1s=1hr
+    time_scale: float = Field(default=0, ge=0)      # 0=max speed, 0.5=demo, 3600=1s=1hr
     scenarios: bool = True                          # Enable scripted events
